+++ plugins/postcss-is-pseudo-class/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased (major)
 
 - Updated: Support for Node v18+ (major).
+- Updated [`@csstools/selector-specificity`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity) to [`4.0.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity/CHANGELOG.md#400) (major)
 
 ### 4.0.8