+++ plugins/postcss-is-pseudo-class/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Make sorting compound selectors forwards compatible with potential future changes in `postcss-selector-parser`
 - Updated: `postcss-selector-parser`
+- Updated [`@csstools/selector-specificity`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity) to [`5.0.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity/CHANGELOG.md#500) (major)
 
 ### 5.0.0